fear(aes) add aes gcm-mode support
Adds AES-GCM mode support in AES cipher. This is an AEAD construct
that is preferred over the default CBC mode that is a subject
to padding oracle. The AES-GCM will be the default on `4.0` when
that is released, but right now the default is not changed for backward
compatibility reasons.