fix(*): hkdf is not approved by FIPS, use PBKDF2 instead on FIPS-mode
### Summary

It turned out that HDKF is only allowed by FIPS in TLS context. This changes
the crypto in this library to use PBKDF2 with SHA-256 using a single iteration
in places where it used HDKF before when in FIPS-mode. This change does not
affect non-fips. Effectively it makes cookies created on FIPS mode incompatible
with cookies created on non-FIPS mode, which is most likely a good thing.